requirements.txt
# AutoGen AG2 - Multi-agent framework
pyautogen>=0.2.0

# Web UI Frameworks
streamlit>=1.28.0
gradio>=4.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# HTTP requests for Jan.ai API
requests>=2.31.0

# Utilities
colorama>=0.4.6
